[Feat] 컨버터 분리
